Added Admin Apis  (#74)

+++ src/main/java/com/bank/webApplication/Services/OTPService.java
@@ -11,6 +11,7 @@ import lombok.extern.slf4j.Slf4j;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.scheduling.annotation.Scheduled;
 import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Transactional;
 
 import java.util.Date;
 import java.util.UUID;
@@ -28,26 +29,26 @@ public class OTPService {
     private OTPRepository otpRepository;
 
     private boolean isExpired(Date expirationTime) {
-        log.info("[OTPService] isExpired SUCCESS");
+        log.info("[OTPService]Entered into isExpired");
         return expirationTime.before(new Date());
     }
 
 
     public UUID sendOTP(String email, UserEntity user) {
-        log.info("[OTPService] sendOTP entered SUCCESS");
+        log.info("[OTPService] Entered into sendOTP");
         Integer otp = otpGenerator.generateOtp();
 
         MailBodyDTO mailBodyDTO = MailBodyDTO.builder()
                 .to(email)
                 .subject("OTP for Forgot Password Request")
-                .text("Welcome to the Banking Application .You have forgot your Password :( . This is the OTP for your forgot Password Request. Please do not forget next time :|" + otp).build();
+                .text("Welcome to the Banking Application .\n  You have forgot your Password :( . This is the OTP for your forgot Password Request :  " + otp + "\n . Please do not forget next time. ").build();
 
         OTPEntity otpEntity = otpRepository.findByUser(user)
                 .orElse(new OTPEntity());
-
-        otpEntity.setOtp(otp);
-        otpEntity.setExpirationTime(new Date(System.currentTimeMillis() + 60 * 1000));
         otpEntity.setUser(user);
+        otpEntity.setOtp(otp);
+        otpEntity.setExpirationTime(new Date(System.currentTimeMillis() + 240 * 1000));
+
 
         emailService.sendMessage(mailBodyDTO);
         OTPService.log.info("[OTPService] sendOTP SUCCESS ");
@@ -74,9 +75,15 @@ public class OTPService {
         return false;
     }
 
-    @Scheduled(fixedRate = 240_000)
+    @Scheduled(fixedRate = 600_000)
+    @Transactional
     public void cleanupExpiredOTP() {
-        otpRepository.deleteByExpirationTimeBefore(new Date());
-        log.info("[OTPService] cleanupExpiredOTP SUCCESS");
+        if (! otpRepository.findAll().isEmpty()) {
+            otpRepository.deleteByExpirationTimeBefore(new Date());
+            log.info("[OTP Service] cleanupExpiredOTP SUCCESS");
+        }
+        else {
+            log.info("[OTP Service]No OTPs found, skipping cleanup.");
+        }
     }
 }
